Feature header check and others. Bug fixes

+++ newsup.pl
@@ -27,30 +27,112 @@ use Config::Tiny;
 use File::Find;
 use File::Basename;
 use Data::Dumper;
-use Carp;
+use Carp qw/croak carp/;
 use Time::HiRes qw/gettimeofday/;
 use POSIX qw/ceil/;
 use Compress::Zlib;
-use Data::Dumper;
 use IO::Socket::INET;
 use IO::Socket::SSL;# qw(debug3);
 use File::Path qw(remove_tree);
 
+use Inline C => <<'C_CODE';
+//Thank you Tomas Novysedlak for this piece of code :-)
+char* _yenc_encode_c(unsigned char* data, size_t data_size)
+{
+	const unsigned char maxwidth = 128;
+
+	unsigned char *pointer, *encbuffer;
+	size_t encoded_size = data_size;
+	int column = 0;
+	unsigned char c;
+	int i;
+
+	if (NULL == (encbuffer = malloc(data_size << 1)))
+	{
+		return NULL;
+	}
+	else
+	{
+		pointer = encbuffer;
+	}
+
+	for(i = 0; i < data_size; ++i)
+	{
+		c = data[i];
+		c = (c + 42) & 0xFF;
+
+		switch (c)
+		{
+			case 0 :
+			case '\n' :
+			case '\r' :
+			case '=' :
+			  c += 64;
+			  *(pointer++) = '=';
+			  column++;
+			  encoded_size++;
+			  break;
+
+			case '\t' :
+			case ' ' :
+			  if(!column || column - 1 == maxwidth)
+			  {
+				  column++;
+				  encoded_size++;
+				  *(pointer++) = '=';
+				  c += 64;
+			  }
+			  break;
+
+			case '.' :
+			  if(!column)
+			  {
+				  column++;
+				  encoded_size++;
+				  *(pointer++) = '=';
+				  c += 64;
+			  }
+			  break;
+		}
+
+		*(pointer++) = c;
+		column++;
+
+		if(column >= maxwidth)
+		{
+			column = 0;
+			*(pointer++) = '\r';
+			*(pointer++) = '\n';
+			encoded_size += 2;
+		}
+	}
+
+	*pointer = 0;
+	encoded_size++;
+	return realloc(encbuffer, encoded_size);
+}
+
+C_CODE
+
+
+
 
 $|=1;
 
-#YENC variables used for yenc'ing
+#YENC variables used for yenc'ing in Perl
 my $YENC_NNTP_LINESIZE=128;
 my $NNTP_MAX_UPLOAD_SIZE=750*1024;
 my @YENC_CHAR_MAP = map{
 	my $char = ($_+42)%256;
 	($char == 0 || $char == 10 || $char == 13 || $char == 61) ? '='.chr($char+64) : chr($char);
 
-	} (0..0xffff);
+      } (0..0xffff);
 my %FIRST_TRANSLATION_TABLE=("\x09", "=I", "\x20", "=`", "\x2e","=n");
 my %LAST_TRANSLATION_TABLE=("\x09", "=I", "\x32","=r");
+# END of the yenc variables
+my $STATS_FILE = "./.stats.newsup";
 
-
+my %MESSAGE_IDS=();
 
 #Returns a bunch of options that it will be used on the upload. Options passed through command line have precedence over
 #options on the config file
@@ -157,7 +239,7 @@ sub _parse_command_line{
 
 
     $tempDir = $config->{generic}{tempDir} if exists $config->{generic}{tempDir};
-    croak "Please define a valid temporary dir in the configuration file" if !defined $tempDir || !-d $tempDir;
+    croak "Please define a valid temporary dir in the configuration file" if (!defined $tempDir || !-d $tempDir);
 
     
     if ( @newsGroups == 0) {
@@ -186,11 +268,10 @@ sub _get_files_to_upload{
   my $filesToUploadRef = shift;
 
   my $tempFilesRef=[];
-  for (@$filesToUploadRef) {
-    my $dir = $_;
+  for my $dir (@$filesToUploadRef) {
+
     find(sub{
-	   
-	   if (-f) {
+	   if (-f $_) {
 	     my $newName = $File::Find::name;
 	     push @$tempFilesRef, $newName;
 	     
@@ -217,33 +298,98 @@ sub main{
   $size += -s $_ for @$files;
   $size /=1024;
   
-  my $headers="From: $from\r\nNewsgroups: ".join('',@$newsGroupsRef)."\r\n";
+  my $headers="From: $from\r\nNewsgroups: ".join(',',@$newsGroupsRef)."\r\n";
   my $init=time();
-  #  my $searchFolders = _launch_yenc_processes($connections,$files, $tempDir, $headers, $commentsRef);
-  my $searchFolders = _launch_yenc_processes(8,$files, $tempDir, $headers, $commentsRef);
+  my $searchFolders = _launch_yenc_processes(_get_available_cpus() ,$files, $tempDir, $headers, $commentsRef);
   say "YENC Conversion finished! Starting upload!";
-  
-  _launch_upload_processes($server, $port, $username, $userpasswd, 8, $searchFolders);
+
+  _launch_upload_processes($server, $port, $username, $userpasswd, $connections, $searchFolders);
+
+  if ($headerCheck) {
+    _launch_header_check($headerCheckServer, $headerCheckPort, $headerCheckUsername, $headerCheckPassword, $headerCheckSleep,
+			 $connections, $newsGroupsRef->[0],$searchFolders);
+    _launch_upload_processes($server, $port, $username, $userpasswd, $connections, $searchFolders);
+  }
   my $time = time()-$init;
 
-  say "Transfered ".int($size/1024)."MB in ".int($time/60)."m ".($time%60)."s. Speed: [".int($size/$time)." KBytes/Sec]";
-  _create_nzb($searchFolders, $newsGroupsRef);
-  remove_tree(@$searchFolders,1,0);
+  $time=1 if($time==0);
   
+  if (_look_for_failed_uploads($searchFolders)) {
+    say "Transfered ".int($size/1024)."MB in ".int($time/60)."m ".($time%60)."s. Speed: [".int($size/$time)." KBytes/Sec]";
+    _create_nzb($nzbName, $searchFolders, $newsGroupsRef);
+    remove_tree(@$searchFolders,1,0);  
+  }
+}
+
+sub _launch_header_check{
+  my ($headerCheckServer, $headerCheckPort, $headerCheckUsername, $headerCheckPassword,
+      $headerCheckSleep, $connections, $newsgroup, $searchFolders) = @_;
+
+  sleep($headerCheckSleep);
+  my $socket = _create_socket($headerCheckServer, $headerCheckPort);
+  croak "Unable to login. Please check the credentials" if _authenticate($socket, $headerCheckUsername, $headerCheckPassword) == 1;
+
+  find(sub{
+	 if (-f $File::Find::name && $_ =~ /\.uploaded$/ ) {
+
+	   _header_check($socket, $newsgroup, $_, $File::Find::name);
+	   
+	 }
+       },@$searchFolders);
+  _logout($socket);
+  
+}
+
+sub _header_check{
+  my ($socket, $newsgroup,$id, $file) = @_;
+
+  croak "Unable to print to socket" if (_print_to_socket($socket, "GROUP $newsgroup\r\n")!=0);
+  my $output = _read_from_socket($socket);
+
+  if ($output =~ /^211\s/) {
+    $id =~ s/\.uploaded$//;
+    croak "Unable to print to socket" if (_print_to_socket($socket, "head <$id>\r\n")!=0);
+    
+    if ($output =~ /^221\s$/){
+      while ( _read_from_socket($socket)){ last if $_ =~ /^\.$/};
+    }else {
+      (my $newName = $file)=~ s/\.uploaded$//;
+      rename($file, $newName);
+    }
+  }
+
+}
+
+sub _look_for_failed_uploads{
+  my @folders = @{shift(@_)};
+  my $success = 1;
+
+  find(sub{
+	 
+	 if (-e $File::Find::name && $_ =~ /\.failed$|\.newsup$/ ) {
+	   my @fileData = fileparse($File::Find::name);
+	   say "Article ".$fileData[0]." failed!";
+	   $success=0;
+
+	 }
+	 
+       },@folders) if @folders;
+
+  return $success;
   
 }
 
 sub _create_nzb{
-  my ($folders, $newsGroups)=@_;
-
+  my ($nzbName, $folders, $newsGroups)=@_;
+  $nzbName='newsup.nzb' if (!defined $nzbName);
+  
   my $subjectRegexp = qr/Subject: .*"(.*)" yenc\s?\((\d+)\/\d+\)/;
   my $sizeRegexp = qr/size=(\d+)/;
   my %nzb=();
   find(sub{
-
 	 
 	 if (-e $File::Find::name && !-d $File::Find::name && $_ =~ /uploaded$/) {
-	   open my $ifh, '<', $File::Find::name;
+	   open my $ifh, '<:raw', $File::Find::name;
 
 	   while ((my $line = <$ifh>)) {
 	     if ($line =~ /$subjectRegexp/) {
@@ -258,22 +404,19 @@ sub _create_nzb{
 	 
        },@$folders);
 
-  open my $ofh, '>', 'newsup.nzb';
-  print $ofh "<nzb xmlns=\"http://www.newzbin.com/DTD/2003/nzb\">\r\n";
+  open my $ofh, '>', $nzbName;
+  print $ofh "<?xml version=\"1.0\" encoding=\"iso-8859-1\" ?>\n<nzb xmlns=\"http://www.newzbin.com/DTD/2003/nzb\">\n";
   my $date = time();
   for my $k (keys %nzb) {
-    print $ofh "<file poster=\"newsup\" date=\"$date\" subject=\"$k\">\r\n";
-    print $ofh "<groups>\r\n";
-    print $ofh "<group>$_</group>\r\n" for @$newsGroups;
-    print $ofh "</groups>\r\n";
-    print $ofh "<segments>\r\n";
+    print $ofh "<file poster=\"newsup\" date=\"$date\" subject=\"&quot;$k&quot;\">\n<groups>\n";
+    print $ofh "<group>$_</group>\n" for @$newsGroups;
+    print $ofh "</groups>\n<segments>\n";
     for (sort {$a->[1] <=> $b->[1]} @{$nzb{$k}}) {
       my $segment = $_->[0];
       my $segmentNumber = $_->[1];
-      print $ofh "<segment bytes=\"$NNTP_MAX_UPLOAD_SIZE\" number=\"$segmentNumber\">$segment</segment>\r\n";
+      print $ofh "<segment bytes=\"$NNTP_MAX_UPLOAD_SIZE\" number=\"$segmentNumber\">$segment</segment>\n";
     }
-    print $ofh "</segments>\r\n";
-    print $ofh "</file>\r\n";
+    print $ofh "</segments>\n</file>\n";
   }
   
   print $ofh "</nzb>";
@@ -311,50 +454,72 @@ sub _launch_upload{
   }
 
   my $socket = _create_socket($server, $port);
-  croak "Unable to login. Please check the credentials" if _authenticate($socket, $user, $password) == -1;
+  croak "Unable to login. Please check the credentials" if _authenticate($socket, $user, $password) == 1;
   
-
   find(sub{
-
-	 if (-e $File::Find::name && !-d $File::Find::name && $_ =~ /newsup$/) {
+	 
+	 if (-e $File::Find::name && -f $File::Find::name && $_ =~ /newsup$/) {
 	   if (rename $File::Find::name, $File::Find::name.".lock") {
-	     _post_file($socket, $File::Find::name.".lock");
-	     
-	     rename $File::Find::name.'.lock', $File::Find::name.".uploaded";
-	     
-	   } 
-	   
+	     eval{
+	       _post_file($socket, $File::Find::name.".lock", $user, $password);
+	     };
+	     if (@!) {
+	       rename $File::Find::name.'.lock', $File::Find::name.".failed";
+	       exit 0;
+	     }else {
+	       rename $File::Find::name.'.lock', $File::Find::name.".uploaded";
+	     }
+
+	   }
 	 }
 	 
        },@$folders);
   
   _logout ($socket);
-  
+
   exit 0;
 }
 
 sub _post_file{
-  my ($socket,  $file) = @_;
+  my ($socket,  $file, $user, $password) = @_;
 
   #_read_from_socket($socket,$select);
-  croak "Unable to print to socket" if (_print_to_socket ($socket, "POST\r\n") == -1);
+  croak "Unable to print to socket" if (_print_to_socket ($socket, "POST\r\n") != 0);
   my $output = _read_from_socket($socket);
 
   croak "Unable to POST (no 340): $output" if($output !~ /340/);
   my $data;
-  {
-    open my $ifh, '<', $file or die "Cannot open the file '$file': $!";
+  eval{
+    open my $ifh, '<:raw', $file or die "Cannot open the file '$file': $!";
     local $/=undef;
     my $data = <$ifh>;
     _print_to_socket ($socket, $data);
 
     close $ifh;
-  }
-  
+  };
+  croak "Error uploading: $!" if @!;
   
   $output = _read_from_socket($socket);
+  $output =~ s/\r\n//;
+
+  if ($output =~ /^400\s/) {#Session expired
+    say "Session expired! Re-authenticating: $output";
+    _logout($socket);
+    $socket = _create_socket($socket->peerhost(), $socket->peerport());
+    croak "Unable to login. Please check the credentials" if _authenticate($socket, $user, $password) == 1;
+    say "Repost successull?: ". (_post_file($socket, $file, $user, $password)==0);
+  }else {
+    if($output !~ /^240/){
+      carp "Error posting article: $output";
+      (my $new_file_name = $file) =~ s/\.lock$/.failed/;
+      rename($file, $new_file_name);
+
+      return 1;
+    }
+  }
+
+  return 0;
   
-  croak "Error posting article: $output " if($output!~ /240/);
 }
 
 
@@ -364,12 +529,11 @@ sub _logout{
   shutdown $socket, 2;  
 }
 
-
 sub _launch_yenc_processes{
   my ($processes, $files, $tmpDir, $headers, $comments) = @_;
 
   my $filesPerProcesses = _split_files_per_process($processes, $files);
-
+  
   my @processes = ();
   for (0..$processes-1) {
     push @processes, _create_yenc_articles($filesPerProcesses->[$_], $tmpDir, $headers, $comments);
@@ -387,7 +551,7 @@ sub _launch_yenc_processes{
 
 sub _create_yenc_articles{
   my ($files, $tmpDir, $headers, $commentsRef) = @_;
-
+  
   my $pid;
   
   unless (defined($pid = fork())) {
@@ -398,45 +562,53 @@ sub _create_yenc_articles{
     return $pid; # I'm the parent
   }
 
-  
+  my ($seed, $usec) = gettimeofday();
+
   for my $fileData (@$files) {
-    open my $ifh , '<:bytes', $fileData->[1];
-    
     my $fileName = (fileparse($fileData->[1]))[0];
     my $dir ="$tmpDir/${fileName}_yenc";
-    mkdir $dir or die "Unable to create directory '$dir': $!";
-    my $bytes;
-    my $part = 1;
-    my $totalParts = ceil((-s $fileData->[1])/$NNTP_MAX_UPLOAD_SIZE);
-    my $startPosition=1;
-    my $subject = '['.$fileData->[0].'] - "'.$fileName.'"';
-    if (defined $commentsRef) {
-      $subject = $commentsRef->[0].' '.$subject if scalar(@$commentsRef) > 0;
-      $subject = $subject . ' ['.$commentsRef->[1].']' if scalar(@$commentsRef) > 1;
-    }
-
-
-
     
-    while ((my $readedSize = read($ifh, $bytes, $NNTP_MAX_UPLOAD_SIZE)) ) {
-      my $messageID = _get_message_id();
-      open my $ofh, '>:raw',$dir.'/'.$messageID;
-      my $yencData=_yenc_encode($bytes);
-      print $ofh $headers,
-	"Subject: ",$subject,' yenc (',$part,'/',$totalParts,")\r\n",
-	"Message-ID: <",$messageID,">\r\n",
-	"\r\n",
-	"=ybegin part=",$part," total=",$totalParts, " line=",$YENC_NNTP_LINESIZE, " size=", $readedSize, " name=",$fileName,"\r\n",
-	"=ypart begin=",$startPosition," end=",tell $ifh,"\r\n",
-	${$yencData->[1]},
-	"\r\n=yend size=",$readedSize," pcrc32=",$yencData->[0],"\r\n.\r\n";
-
-      $startPosition=1+tell $ifh;
-      $part++;
+    if (!-e $dir ) {
+      open my $ifh , '<', $fileData->[1];
       
-      close $ofh;
+      my $fileSize = -s $fileData->[1];
+    
+      mkdir $dir or die "Unable to create directory '$dir': $!";
+      my $bytes;
+      my $part = 1;
+      my $totalParts = ceil((-s $fileData->[1])/$NNTP_MAX_UPLOAD_SIZE);
+      my $startPosition=1;
+      my $subject = '['.$fileData->[0].'] - "'.$fileName.'"';
+      if (defined $commentsRef) {
+	$subject = $commentsRef->[0].' '.$subject if scalar(@$commentsRef) > 0;
+	$subject = $subject . ' ['.$commentsRef->[1].']' if scalar(@$commentsRef) > 1;
+      }
+      
+      while ((my $readedSize = read($ifh, $bytes, $NNTP_MAX_UPLOAD_SIZE)) ) {
+	my $messageID = $part.$fileData->[2];
+	open my $ofh, '>:raw',$dir.'/'.$messageID;
+	my $yencData=_yenc_encode_c($bytes, $readedSize);
+	print $ofh $headers,
+	  "Subject: ",$subject,' yenc (',$part,'/',$totalParts,")\r\n",
+	  "Message-ID: <",$messageID,">\r\n",
+	  "\r\n",
+	  "=ybegin part=",$part," total=",$totalParts, " line=",$YENC_NNTP_LINESIZE, " size=", $fileSize, " name=",$fileName,"\r\n",
+	  "=ypart begin=",$startPosition," end=",tell $ifh,"\r\n",
+	  #_yenc_encode_c($YENC_NNTP_LINESIZE,$bytes),
+	  $yencData,
+	  #"\r\n=yend size=",$readedSize," pcrc32=",$yencData->[0],"\r\n.\r\n";
+	  "\r\n=yend size=",$readedSize," pcrc32=",sprintf("%x",crc32 ($bytes)),"\r\n.\r\n";
+	
+	$startPosition=1+tell $ifh;
+	$part++;
+	
+	close $ofh;
+      }
+      close $ifh;
+    }else {
+      say "Found old folder from a failed(?) upload. Using it instead: $dir";
     }
-    close $ifh;
+
   }
 
   exit 0;
@@ -450,7 +622,7 @@ sub _split_files_per_process{
   
   my $i=0;
   foreach my $elem (sort @$files) {
-    push @{ $parts[$i++ % $processes] }, ["$i/$total",$elem];
+    push @{ $parts[$i++ % $processes] }, ["$i/$total",$elem, _get_message_id() ];
   }
 
   return \@parts;
@@ -463,7 +635,15 @@ sub _get_message_id{
   my $time = _encode_base36("$s$usec");
   my $randomness = _encode_base36(rand("$s$usec"));
 
-  return "$s$usec.$randomness\@$time.newsup";
+  my $mid = "$s$usec.$randomness\@$time.newsup";
+
+  if (!exists $MESSAGE_IDS{$mid}) {
+    $MESSAGE_IDS{$mid}=1;
+    return $mid;
+  }else {
+    return _get_message_id();
+  }
+
 }
 
 sub _encode_base36 {
@@ -477,7 +657,23 @@ sub _encode_base36 {
   return $b36 || '0';
 }
 
+sub _get_available_cpus{
 
+    my $yenc_processes = 2;
+    if($^O =~ /MSWin32|cygwin/){
+      $yenc_processes = $ENV{"NUMBER_OF_PROCESSORS"};
+    }elsif($^O =~ /linux/){
+      open my $ifh, '<', '/proc/cpuinfo';
+      while(<$ifh>){
+	$yenc_processes+=1 if(/processor/);
+      }
+      close $ifh;
+    }
+    elsif($^O =~ /hpux/){
+      $yenc_processes = scalar(my @l = qx/ioscan -k -C processor/)-2;
+    }
+    return $yenc_processes;
+}
 
 
 sub _yenc_encode{
@@ -488,7 +684,7 @@ sub _yenc_encode{
 
     my $char= $YENC_CHAR_MAP[$hexChar];
     
-    if($char =~ /=/){
+    if($char =~ /^=.$/){
       $column++;
     }
     elsif($column==0 && $FIRST_TRANSLATION_TABLE{$char}){
@@ -520,7 +716,7 @@ sub _read_from_socket{
     $socket->sysread($buffer, 1024);
 
     $output .= $buffer;
-    last if $output =~ /\r\n$/;
+    last if $output =~ /\r\n$|^\z/;
   }
 
   return $output;
@@ -531,7 +727,7 @@ sub _print_to_socket{
 
   print $socket $args;
 
-  return 1;
+  return 0;
 }
 
 
@@ -539,17 +735,21 @@ sub _authenticate{
   my ($socket,  $user, $password) = @_;
 
   my $output = _read_from_socket $socket;
-  croak "Unable to print to socket" if (_print_to_socket ($socket, "authinfo user $user\r\n") == -1);
+  croak "Unable to print to socket" if (_print_to_socket ($socket, "authinfo user $user\r\n") != 0);
 
   $output =  _read_from_socket $socket;
   croak $output if $output !~ /381/;
 
-  croak "Unable to print to socket" if (_print_to_socket ($socket, "authinfo pass $password\r\n") == -1);
+  croak "Unable to print to socket" if (_print_to_socket ($socket, "authinfo pass $password\r\n") != 0);
   
   $output =  _read_from_socket $socket;
-  
-  croak $output if $output !~ /281/;
 
+  if ($output !~ /281/){
+    carp $output;
+    return 1;
+  }
+  0;
+  
 }
 
 sub _create_socket{
@@ -564,6 +764,7 @@ sub _create_socket{
 				   SSL_verify_mode=>SSL_VERIFY_NONE,
 				   SSL_version=>'TLSv1',
 				   Blocking => 1,
+				   Timout=> 20,
 				   #SSL_version=>'TLSv1_2',
 				   #SSL_cipher_list=>'DHE-RSA-AES128-SHA',
 				   SSL_ca_path=>'/etc/ssl/certs',
@@ -574,6 +775,7 @@ sub _create_socket{
 				     PeerPort => $port,
 				     Blocking => 1,
 				     Proto => 'tcp',
+				     Timeout => 20,
 				    ) or die "ERROR in Socket Creation : $!\n";
   }
   
@@ -583,4 +785,3 @@ sub _create_socket{
 }
 
 main();
-
